+++ components/catalogs/back/tests/Integration/Infrastructure/Controller/Public/DeleteProductMappingSchemaActionTest.php
@@ -1,0 +1,182 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Akeneo\Catalogs\Test\Integration\Infrastructure\Controller\Public;
+
+use Akeneo\Catalogs\Domain\Catalog;
+use Akeneo\Catalogs\Infrastructure\Persistence\Catalog\GetCatalogQuery;
+use Akeneo\Catalogs\ServiceAPI\Command\CreateCatalogCommand;
+use Akeneo\Catalogs\ServiceAPI\Command\UpdateProductMappingSchemaCommand;
+use Akeneo\Catalogs\ServiceAPI\Exception\ProductSchemaMappingNotFoundException as ServiceApiProductSchemaMappingNotFoundException;
+use Akeneo\Catalogs\ServiceAPI\Messenger\CommandBus;
+use Akeneo\Catalogs\ServiceAPI\Messenger\QueryBus;
+use Akeneo\Catalogs\ServiceAPI\Query\GetProductMappingSchemaQuery;
+use Akeneo\Catalogs\Test\Integration\IntegrationTestCase;
+use PHPUnit\Framework\Assert;
+use Symfony\Bundle\FrameworkBundle\KernelBrowser;
+
+/**
+ * @copyright 2022 Akeneo SAS (http://www.akeneo.com)
+ * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ *
+ * @covers \Akeneo\Catalogs\Infrastructure\Controller\Public\DeleteProductMappingSchemaAction
+ * @covers \Akeneo\Catalogs\Application\Handler\DeleteProductMappingSchemaHandler
+ */
+class DeleteProductMappingSchemaActionTest extends IntegrationTestCase
+{
+    private ?KernelBrowser $client = null;
+    private ?CommandBus $commandBus;
+    private ?QueryBus $queryBus;
+
+    protected function setUp(): void
+    {
+        parent::setUp();
+
+        $this->commandBus = self::getContainer()->get(CommandBus::class);
+        $this->queryBus = self::getContainer()->get(QueryBus::class);
+
+        $this->purgeDataAndLoadMinimalCatalog();
+    }
+
+    public function testItUpdatesTheCatalogProductMappingSchema(): void
+    {
+        $this->client = $this->getAuthenticatedPublicApiClient([
+            'write_catalogs',
+        ]);
+        $this->createCatalog('db1079b6-f397-4a6a-bae4-8658e64ad47c', 'Store US', 'shopifi');
+        $this->commandBus->execute(new UpdateProductMappingSchemaCommand(
+            'db1079b6-f397-4a6a-bae4-8658e64ad47c',
+            \json_decode($this->getValidSchemaData(), false, 512, JSON_THROW_ON_ERROR),
+        ));
+        $this->setCatalogProductMapping('db1079b6-f397-4a6a-bae4-8658e64ad47c', [
+            'name' => [
+                'attribute' => 'title',
+                'scope' => 'ecommerce',
+                'locale' => 'en_US',
+            ],
+        ]);
+
+        $this->client->request(
+            'DELETE',
+            '/api/rest/v1/catalogs/db1079b6-f397-4a6a-bae4-8658e64ad47c/mapping-schemas/product',
+            [],
+            [],
+            [
+                'CONTENT_TYPE' => 'application/json',
+            ],
+        );
+
+        $response = $this->client->getResponse();
+
+        Assert::assertEquals(204, $response->getStatusCode());
+
+        $this->expectException(ServiceApiProductSchemaMappingNotFoundException::class);
+        $this->queryBus->execute(
+            new GetProductMappingSchemaQuery('db1079b6-f397-4a6a-bae4-8658e64ad47c')
+        );
+
+        /** @var Catalog $catalog */
+        $catalog = self::getContainer()->get(GetCatalogQuery::class)->execute('db1079b6-f397-4a6a-bae4-8658e64ad47c');
+        Assert::assertJsonStringEqualsJsonString(
+            '[]',
+            \json_encode($catalog->getProductMapping(), JSON_THROW_ON_ERROR)
+        );
+    }
+
+    public function testItReturnsForbiddenWhenMissingPermissions(): void
+    {
+        $this->client = $this->getAuthenticatedPublicApiClient([]);
+        $this->commandBus->execute(new CreateCatalogCommand(
+            'db1079b6-f397-4a6a-bae4-8658e64ad47c',
+            'Store US',
+            'shopifi',
+        ));
+
+        $this->client->request(
+            'DELETE',
+            '/api/rest/v1/catalogs/db1079b6-f397-4a6a-bae4-8658e64ad47c/mapping-schemas/product',
+            [],
+            [],
+            [
+                'CONTENT_TYPE' => 'application/json',
+            ],
+        );
+
+        $response = $this->client->getResponse();
+
+        Assert::assertEquals(403, $response->getStatusCode());
+    }
+
+    public function testItReturnsNotFoundWhenCatalogDoesNotExist(): void
+    {
+        $this->client = $this->getAuthenticatedPublicApiClient([
+            'write_catalogs',
+        ]);
+
+        $this->client->request(
+            'DELETE',
+            '/api/rest/v1/catalogs/db1079b6-f397-4a6a-bae4-8658e64ad47c/mapping-schemas/product',
+            [],
+            [],
+            [
+                'CONTENT_TYPE' => 'application/json',
+            ],
+        );
+
+        $response = $this->client->getResponse();
+
+        Assert::assertEquals(404, $response->getStatusCode());
+    }
+
+    public function testItReturnsNotFoundWhenCatalogDoesNotBelongToCurrentUser(): void
+    {
+        $this->client = $this->getAuthenticatedPublicApiClient([
+            'write_catalogs',
+        ]);
+        $this->createUser('magendo');
+        $this->commandBus->execute(new CreateCatalogCommand(
+            'db1079b6-f397-4a6a-bae4-8658e64ad47c',
+            'Store US',
+            'magendo',
+        ));
+
+        $this->client->request(
+            'DELETE',
+            '/api/rest/v1/catalogs/db1079b6-f397-4a6a-bae4-8658e64ad47c/mapping-schemas/product',
+            [],
+            [],
+            [
+                'CONTENT_TYPE' => 'application/json',
+            ],
+        );
+
+        $response = $this->client->getResponse();
+
+        Assert::assertEquals(404, $response->getStatusCode());
+    }
+
+    private function getValidSchemaData(): string
+    {
+        return <<<'JSON_WRAP'
+        {
+          "$id": "https://example.com/product",
+          "$schema": "https://api.akeneo.com/mapping/product/0.0.1/schema",
+          "$comment": "My first schema !",
+          "title": "Product Mapping",
+          "description": "JSON Schema describing the structure of products expected by our application",
+          "type": "object",
+          "properties": {
+            "name": {
+              "type": "string"
+            },
+            "body_html": {
+              "title": "Description",
+              "description": "Product description in raw HTML",
+              "type": "string"
+            }
+          }
+        }
+        JSON_WRAP;
+    }
+}