+++ components/identifier-generator/back/tests/.php_cd.php
@@ -43,11 +43,13 @@ $rules = [
             'Symfony\Component\DependencyInjection\Loader\YamlFileLoader',
             'Symfony\Component\HttpKernel\Bundle\Bundle',
             'Symfony\Component\HttpKernel\DependencyInjection\Extension',
+            'Symfony\Component\HttpKernel\Exception\BadRequestHttpException',
             'Symfony\Component\EventDispatcher\EventSubscriberInterface',
             'Symfony\Component\HttpFoundation',
             'Symfony\Component\Validator',
-            'Doctrine\DBAL\Exception',
-            'Doctrine\DBAL\Connection',
+
+            'Doctrine\DBAL',
+
             'Ramsey\Uuid\Uuid',
             'Webmozart\Assert\Assert',
         ]