+++ components/identifier-generator/back/tests/EndToEnd/Infrastructure/Controller/GetIdentifierAttributesControllerEndToEnd.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Akeneo\Test\Pim\Automation\IdentifierGenerator\EndToEnd\Infrastructure\Controller;
 
 use Akeneo\Test\Integration\Configuration;