+++ components/identifier-generator/front/src/feature/components/Styled.tsx
@@ -1,0 +1,20 @@
+import styled from 'styled-components';
+
+const FormContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+  margin-top: 20px;
+  margin-bottom: 20px;
+`;
+
+const TabContainer = styled.div`
+  padding: 20px 40px;
+`;
+
+const Styled = {
+  FormContainer,
+  TabContainer,
+};
+
+export {Styled};