+++ components/identifier-generator/front/src/feature/hooks/useIdentifierAttributes.ts
@@ -1,9 +1,16 @@
 import {useQuery} from 'react-query';
-import {FlattenAttribute} from '../models/flatten-attribute';
+import {FlattenAttribute} from '../models';
+import {useRouter} from '@akeneo-pim-community/shared';
 
-const useIdentifierAttributes = () => {
+const useIdentifierAttributes: () => {
+  data?: FlattenAttribute[];
+  error: Error | null;
+  isSuccess: boolean;
+} = () => {
+  const router = useRouter();
+
   const getIdentifierAttributes = async () => {
-    return fetch('/identifier-generator/identifier-attributes', {
+    return fetch(router.generate('akeneo_identifier_generator_get_identifier_attributes'), {
       method: 'GET',
       headers: [['X-Requested-With', 'XMLHttpRequest']],
     }).then(res => {
@@ -12,17 +19,12 @@ const useIdentifierAttributes = () => {
     });
   };
 
-  const {error, data} = useQuery<FlattenAttribute[], Error, FlattenAttribute[]>(
+  const {error, data, isSuccess} = useQuery<FlattenAttribute[], Error, FlattenAttribute[]>(
     'getIdentifierAttributes',
-    getIdentifierAttributes,
-    {
-      keepPreviousData: true,
-      refetchOnWindowFocus: false,
-      retry: false,
-    }
+    getIdentifierAttributes
   );
 
-  return {data, error};
+  return {data, error, isSuccess};
 };
 
 export {useIdentifierAttributes};