+++ components/identifier-generator/front/src/feature/models/properties/autoNumber.ts
@@ -1,0 +1,7 @@
+import {PROPERTY_NAMES} from '../structure';
+
+type AutoNumber = {
+  type: PROPERTY_NAMES.AUTO_NUMBER;
+};
+
+export type {AutoNumber};