+++ components/identifier-generator/front/src/validators/validateConditions.ts
@@ -1,8 +1,0 @@
-import {Validator} from './Validator';
-import {Conditions} from '../models';
-
-const validateConditions: Validator<Conditions> = (_conditions, _path) => {
-  return [];
-};
-
-export {validateConditions};