+++ src/Akeneo/Connectivity/Connection/back/tests/Unit/spec/Infrastructure/Apps/AsymmetricKeysGeneratorSpec.php
@@ -1,4 +1,5 @@
 <?php
+
 declare(strict_types=1);
 
 namespace spec\Akeneo\Connectivity\Connection\Infrastructure\Apps;
@@ -13,6 +14,11 @@ use PhpSpec\ObjectBehavior;
  */
 class AsymmetricKeysGeneratorSpec extends ObjectBehavior
 {
+    public function let()
+    {
+        $this->beConstructedWith(__DIR__ . '/openssl.cnf');
+    }
+
     public function it_is_an_asymmetric_keys_generator(): void
     {
         $this->shouldHaveType(AsymmetricKeysGenerator::class);