+++ tests/back/Pim/Enrichment/Specification/Component/Product/Normalizer/InternalApi/ProductModelNormalizerSpec.php
@@ -140,7 +140,17 @@ class ProductModelNormalizerSpec extends ObjectBehavior
                 'prices'              => [['data' => 12.5, 'locale' => null, 'scope' => null]],
                 'date'                => [['data' => '2015-01-31', 'locale' => null, 'scope' => null]],
                 'picture'             => [['data' => 'a/b/c/my_picture.jpg', 'locale' => null, 'scope' => null]]
-            ]
+            ],
+            'quantified_associations' => [
+                'set' => [
+                    'products' => [
+                        ['uuid' => '79fc4791-86d6-4d3b-93c5-76b787af9497', 'identifier' => 'a_product', 'quantity' => 3],
+                    ],
+                    'product_models' => [
+                        ['identifier' => 'a_product_model', 'quantity' => 10],
+                    ],
+                ],
+            ],
         ];
 
         $familyVariantNormalized = [
@@ -255,6 +265,16 @@ class ProductModelNormalizerSpec extends ObjectBehavior
                 'family'         => 'tshirts',
                 'categories'     => ['summer'],
                 'values'         => $valuesConverted,
+                'quantified_associations' => [
+                    'set' => [
+                        'products' => [
+                            ['uuid' => '79fc4791-86d6-4d3b-93c5-76b787af9497', 'quantity' => 3],
+                        ],
+                        'product_models' => [
+                            ['identifier' => 'a_product_model', 'quantity' => 10],
+                        ],
+                    ],
+                ],
                 'parent_associations' => null,
                 'meta'           => [
                     'variant_product_completenesses' => [
@@ -334,7 +354,17 @@ class ProductModelNormalizerSpec extends ObjectBehavior
                 'prices'              => [['data' => 12.5, 'locale' => null, 'scope' => null]],
                 'date'                => [['data' => '2015-01-31', 'locale' => null, 'scope' => null]],
                 'picture'             => [['data' => null, 'locale' => null, 'scope' => null]]
-            ]
+            ],
+            'quantified_associations' => [
+                'set' => [
+                    'products' => [
+                        ['uuid' => '79fc4791-86d6-4d3b-93c5-76b787af9497', 'identifier' => 'a_product', 'quantity' => 3],
+                    ],
+                    'product_models' => [
+                        ['identifier' => 'a_product_model', 'quantity' => 10],
+                    ],
+                ],
+            ],
         ];
 
         $familyVariantNormalized = [
@@ -430,6 +460,16 @@ class ProductModelNormalizerSpec extends ObjectBehavior
                 'family'         => 'tshirts',
                 'categories'     => ['summer'],
                 'values'         => $valuesConverted,
+                'quantified_associations' => [
+                    'set' => [
+                        'products' => [
+                            ['uuid' => '79fc4791-86d6-4d3b-93c5-76b787af9497', 'quantity' => 3],
+                        ],
+                        'product_models' => [
+                            ['identifier' => 'a_product_model', 'quantity' => 10],
+                        ],
+                    ],
+                ],
                 'parent_associations' => null,
                 'meta'           => [
                     'variant_product_completenesses' => [
@@ -506,7 +546,17 @@ class ProductModelNormalizerSpec extends ObjectBehavior
                 'prices'              => [['data' => 12.5, 'locale' => null, 'scope' => null]],
                 'date'                => [['data' => '2015-01-31', 'locale' => null, 'scope' => null]],
                 'picture'             => [['data' => 'a/b/c/my_picture.jpg', 'locale' => null, 'scope' => null]]
-            ]
+            ],
+            'quantified_associations' => [
+                'set' => [
+                    'products' => [
+                        ['uuid' => '79fc4791-86d6-4d3b-93c5-76b787af9497', 'identifier' => 'a_product', 'quantity' => 3],
+                    ],
+                    'product_models' => [
+                        ['identifier' => 'a_product_model', 'quantity' => 10],
+                    ],
+                ],
+            ],
         ];
 
         $familyVariantNormalized = [
@@ -613,6 +663,16 @@ class ProductModelNormalizerSpec extends ObjectBehavior
                 'family'         => 'tshirts',
                 'categories'     => ['summer'],
                 'values'         => $valuesConverted,
+                'quantified_associations' => [
+                    'set' => [
+                        'products' => [
+                            ['uuid' => '79fc4791-86d6-4d3b-93c5-76b787af9497', 'quantity' => 3],
+                        ],
+                        'product_models' => [
+                            ['identifier' => 'a_product_model', 'quantity' => 10],
+                        ],
+                    ],
+                ],
                 'parent_associations' => null,
                 'meta'           => [
                     'variant_product_completenesses' => [